requirements.txt
django-crispy-forms==1.9.2
django==3.0.8
djangorestframework==3.11.0
pillow==7.2.0
drf-nested-fields==0.9.5